+++ device-bootloaders/src/ServiceProcessorBL2/services/firmware_update.c
@@ -780,6 +780,13 @@ static int32_t dm_svc_firmware_update(void)
         }
     }
 
+    Log_Write(LOG_LEVEL_CRITICAL, "Programming config region\n");
+    if (0 != flash_fs_write_config_region(1 - sp_bl2_data->flash_fs_bl2_info.active_partition))
+    {
+        MESSAGE_ERROR("flash_fs_write_file: failed to write config data \n");
+        return ERROR_FW_UPDATE_WRITE_CFG_REGION;
+    }
+
     verify_end = timer_get_ticks_count();
     Log_Write(LOG_LEVEL_CRITICAL, "[ETFP] All loaded bytes verified OK!\n");
     Log_Write(LOG_LEVEL_CRITICAL, "[ETFP] Executing exit sequence...\n");